fixed issue with page number input not updating in toolbar
